Do not notify non-group members when at-mentioned.
This commit makes it so that at-mentions do not notify users who are not
members of the group.

I've also prevented mentioned users from receiving "new comment"
notifications, since they are already receiving at-mention notifications
for the comment. However, the code is a bit sloppy. It will need to be
refactored with Rob's upcoming event model refactor.