Set user language preference through param[:locale].
This is a bit quick and dirty and is crying out for a refactor but I'm confident it will work and I really want it to be live asap. The idea is that we can announce new languages and say 'click here to use Loomio in Romanian'. also I want to make a footer that has the list of languages across the bottom of every page.