db/seeds.rb: Alter the `vote_age_per_user_per_motion` constraint to be deferrable
This would ideally belong in `schema.rb`. Unfortunately, anything but trivial unique
constraints are not supported there so in the spirit of keeping things simple,
the constraint is made deferrable as part of seeding the database.

While this particular change is not considered best practice, it does allow the database to
be bootstrapped from `schema.rb` alone.