Copy template title to poll title as well
When using a template, the expected behavior is that it will prefill
the template values to the poll so it saves time when creating
similar proposals. Having the value only in the placeholder is not
as useful as copying the value to the title since the user would
need to retype the whole title everytime they use the template.

This is a regression from the previous behavior, so this patch
makes it behave as before.