Start to resolve the problem of removing someone from a group.
We have a kinda dumb system when presence of inviter_id and absence of revoked_at, means a record will grant guest access, for DiscussionReader and Stance.

Because we would like to record who invited who, even when inviting members, we want to store inviter_id on records for anyone added to a poll, group member or guest.

So let's add a guest boolean when we want to grant access to a thread or poll for people who don't belong to the group.. or when the poll or discussion does not have a group.

This means, we do not need to revoke stances and discussion_readers when removing members from a group, unless they're marked as guests.