tidy up the behaviour of redeem and revoke in MembershipService
When redeeming or revoking a user we remove any guest access that may have previously existed.
This way we don't need to add revoked_at to old stances or discussion readers, which is messy when re-adding people to a group after they were previously removed. Specifically, we don't want to add them to something (a poll) they were removed from before they were removed as a member.