Allow admins to manage memberships of any group
When using the API to manage memberships is very inconvenient to
have to add bots to every group and make them admin of the group in
order to be able to manage members programmatically. This generates noise
because for every group people will be able to see the bot users as
members.

Considering that global admins are already able to manage the
application from the admin interface, it would make sense that they can
also have more permissions when using the API.

This patch allows users that are global admins of the application to
be able to manage users of any group when using the API.